Move to metadata store
