+++ src/Microsoft.Health.Dicom.Blob/Microsoft.Health.Dicom.Blob.csproj
@@ -2,7 +2,7 @@
 
   <PropertyGroup>
     <Description>Azure blob storage utilities for Microsoft's DICOMweb APIs.</Description>
-    <TargetFrameworks>netcoreapp3.1;net5.0</TargetFrameworks>
+    <TargetFramework>net5.0</TargetFramework>
   </PropertyGroup>
 
   <ItemGroup>
