+++ src/Microsoft.Health.Dicom.Metadata/Registration/Experimental/IBlobServiceBuilder.cs
